crud/crud_app/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.18
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==6.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
numpy==1.21.6
openpyxl==3.1.2
packaging==23.1
pandas==1.3.5
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
